+++ src/constants/path.ts
@@ -19,7 +19,7 @@ export const PATH = {
   reportCategoryDetail: "/report/category",
 
   // 자산관리의 세부 설정 페이지 (경로명 수정할 계획)
-  savingsGoal: "/management/savings-goal",
+  savingsGoal: "/management/savings-savingGoal",
   savingDetailSetting: "/saving-detail-setting",
   regularDepositWithdrawal: "/management/regular-deposit-withdrawal",
   DetailSetting: "/management/regular-deposit-withdrawal-detail-setting",
